refactor: renamed "kingdom_core" to "kingdom_sdk"
