+++ setup.cfg
@@ -23,4 +23,4 @@ inherit = false
 multi_line_output = 3
 include_trailing_comma = true
 line_length = 79
-known_first_party = "project_name"  # FIXME: Do set this to the correct 'project_name'.
+known_first_party = "kingdom_sdk"
