+++ tests/conftest.py
@@ -1,13 +1,13 @@
 import pytest
 
-import kingdom_core
+import kingdom_sdk
 
 
 @pytest.fixture
 def base_dir():
-    return kingdom_core._get_base_dir()  # noqa
+    return kingdom_sdk._get_base_dir()  # noqa
 
 
 @pytest.fixture
 def src_dir():
-    return kingdom_core._get_src_dir()  # noqa
+    return kingdom_sdk._get_src_dir()  # noqa
